% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicAdjust-package.R
\docType{package}
\name{OmicAdjust-package}
\alias{OmicAdjust-package}
\title{OmicAdjust: pragmatic p-value redistribution (exploratory)}
\description{
Tools for exploring omics p-value distributions by redistributing
p-values relative to an expected null count at alpha.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JackAuty/OmicAdjust}
  \item Report bugs at \url{https://github.com/JackAuty/OmicAdjust/issues}
}

}
\author{
\strong{Maintainer}: Jack Rivers-Auty \email{jackrauty@gmail.com}

}
